+++ Source/Shaders/Materials/StripeMaterial.glsl
@@ -1,5 +1,5 @@
-uniform vec4 lightColor;
-uniform vec4 darkColor;
+uniform vec4 evenColor;
+uniform vec4 oddColor;
 uniform float offset;
 uniform float repeat;
 uniform bool horizontal;
@@ -13,8 +13,8 @@ czm_material czm_getMaterial(czm_materialInput materialInput)
     float value = fract((coord - offset) * (repeat * 0.5));
     float dist = min(value, min(abs(value - 0.5), 1.0 - value));
     
-    vec4 currentColor = mix(lightColor, darkColor, step(0.5, value));
-    vec4 color = czm_antialias(lightColor, darkColor, currentColor, dist);
+    vec4 currentColor = mix(evenColor, oddColor, step(0.5, value));
+    vec4 color = czm_antialias(evenColor, oddColor, currentColor, dist);
     
     material.diffuse = color.rgb;
     material.alpha = color.a;